pyproject.toml
[tool.poetry]
name = "desafio-crud"
version = "0.1.0"
description = ""
authors = ["ferrer <ferrer.morgado@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.*"
certifi = "2024.7.4"
charset-normalizer = "3.3.2"
exceptiongroup = "1.2.2"
idna = "3.7"
iniconfig = "2.0.0"
packaging = "24.1"
pluggy = "1.5.0"
psycopg2-binary = "2.9.9"
python-dotenv = "1.0.1"
requests = "2.32.3"
tomli = "2.0.1"
urllib3 = "2.2.2"



[tool.poetry.group.dev.dependencies]
pytest = "^8.3.1"
pytest-cov = "^5.0.0"
taskipy = "^1.13.0"
ruff = "^0.5.4"
httpx = "^0.27.0"
ignr = "^2.2"

[tool.ruff]
line-length = 79

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'

[tool.taskipy.tasks]
lint = 'ruff check . && ruff check . --diff'
format = 'ruff check . --fix && ruff format .'
run = 'fastapi dev desafio_crud/app.py'
pre_test = 'task lint'
test = 'pytest -s -x --cov=desafio_crud -vv'
post_test = 'coverage html'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
